fix tests (at least on linux / mac)
