+++ common/engine/replayV2.lua
@@ -5,7 +5,7 @@ local ReplayV2 = {}
 function ReplayV2.transform(replay)
   for i = 1, #replay.players do
     if replay.players[i].settings.inputs then
-      replay.players[i].settings.inputs = uncompress_input_string(replay.players[i].settings.inputs)
+      replay.players[i].settings.inputs = Replay.decompressInputString(replay.players[i].settings.inputs)
     end
 
     if replay.players[i].settings.level then
