--- server/csprng_seed.txt
@@ -1,1 +1,0 @@
-2000
