+++ server/server.lua
@@ -1,5 +1,3 @@
-SERVER_MODE = true -- global to know the server is running the process
-
 local socket = require("common.lib.socket")
 local logger = require("common.lib.logger")
 local class = require("common.lib.class")
