+++ serverLauncher.lua
@@ -1,7 +1,8 @@
 -- We must launch the server from the root directory so all the requires are the right path relatively.
 require("client.src.developer") -- Require developer here so we can debug if the debug flag is set
+require("server.server_globals")
 local util = require("common.lib.util")
-util.addToCPath("./common/lib/sqlite/??")
+util.addToCPath("./common/lib/??")
 util.addToCPath("./server/lib/??")
 require("server.tests.ConnectionTests")
